requirements.txt
# ============================================================
# Cài đặt: pip install -r requirements.txt
# ============================================================

# ===== Web App =====
streamlit==1.52.2

# ===== LangChain =====
langchain==0.1.20
langchain-community==0.0.38
langchain-core==1.1.3
langchain-chroma==0.1.0
langchain-huggingface==1.1.0
langchain-text-splitters==0.0.1
langchain-google-genai==1.0.3
langchain-groq==0.3.8
langsmith==0.4.59

# ===== LLM / AI APIs =====
groq==0.37.1
google-generativeai==0.5.4
google-api-python-client==2.187.0

# ===== Vector Store =====
chromadb==0.4.24

# ===== Embeddings & NLP Models =====
sentence-transformers==5.2.0
transformers==4.57.3
tokenizers==0.22.1
huggingface-hub==0.36.0
torch==2.9.1

# ===== Vietnamese NLP =====
underthesea==8.3.0
underthesea-core==1.0.5
nltk==3.9.2

# ===== Machine Learning =====
scikit-learn==1.8.0
datasets==4.4.1
numpy==1.26.4
scipy==1.16.3

# ===== Database =====
SQLAlchemy==1.4.54

# ===== Data & Utilities =====
pandas==2.3.3
python-dotenv==1.2.1
requests==2.32.5
beautifulsoup4==4.14.3
tqdm==4.67.1
pydantic==2.12.5
click==8.3.1
pillow==12.0.0

# ===== Visualization =====
matplotlib==3.10.8
seaborn==0.13.2

# ===== Data Collection (cài thủ công nếu cần chạy script crawl) =====
# loguru          # dùng trong tdtu_client.py
# selenium        # dùng trong tdtu_client.py
# seleniumwire    # dùng trong tdtu_client.py
